Time: 5 ms (61.59%), Space: 16.1 MB (48.39%) - LeetHub

+++ 0034-find-first-and-last-position-of-element-in-sorted-array/0034-find-first-and-last-position-of-element-in-sorted-array.cpp
@@ -19,7 +19,7 @@ public:
         int l = 0, h = nums.size()-1, ans = nums.size();
         
         while(l<=h){
-            long long m = (l+h)/2;
+            int m = (l+h)/2;
             if(nums[m]>=target){
                 ans = m; //maybe the answer or maybe not.
                 h = m-1;
@@ -35,7 +35,7 @@ public:
         int l = 0, h = nums.size()-1, ans = nums.size();
         
         while(l<=h){
-            long long m = (l+h)/2;
+            int m = (l+h)/2;
             if(nums[m]>target){
                 ans = m; //maybe the answer or maybe not.
                 h = m-1;
